# userapi/api/openapi/userapi.yaml
openapi: "3.0.1"
info:
  title: Social Network – User API
  description: Interactions with user profiles
  version: 1.0.0
  contact:
    email: stepan2volkov@yandex.ru
...
paths:
...
  /api/v1/profiles/{username}/follow:
    post:
      summary: Follow user profile
      tags: [ Profile ]
      operationId: FollowProfile
      security:
        - cookieAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            example: test
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
        404:
          description: Profile not found
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session